+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This file contains highlights of what changes on each version of the [Web Storag
 
 ## Version 0.2.0
 - Breaking change: renamed the `DOMStorage` class to `WebStorage`.
+- Breaking change: removed the `LocalStorage` and `SessionStorage` classes.
 
 ## Version 0.1.0
 - Initial release.