+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file contains highlights of what changes on each version of the [Web Storag
 ## Version 0.2.0
 - Breaking change: renamed the `DOMStorage` class to `WebStorage`.
 - Breaking change: removed the `LocalStorage` and `SessionStorage` classes.
+- Updated the dependencies.
 
 ## Version 0.1.0
 - Initial release.